// client/src/features/scores/score.types.ts
export type Score = {
  registrationNumber: string
  math: number | null
  literature: number | null
  foreignLanguage: number | null
  physics: number | null
  chemistry: number | null
  biology: number | null
  history: number | null
  geography: number | null
  civicEducation: number | null
  foreignLanguageCode: string | null
}